requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pyserial==3.5
pydantic==2.9.2